move optimizer to post processor
