--- docs/reference/optimizers.rst
@@ -1,6 +1,0 @@
-Optimizers
-==========
-
-.. automodule:: llama_index.optimization
-   :members:
-   :inherited-members:
