--- llama_index/optimization/__init__.py
@@ -1,7 +1,0 @@
-"""Optimization."""
-
-from llama_index.optimization.optimizer import SentenceEmbeddingOptimizer
-
-__all__ = [
-    "SentenceEmbeddingOptimizer",
-]
